initial package structure; moved external price method to RO
